Fix CI builds: bump Kotlin 1.4.32→1.5.10, fix metadataSourcesJar duplicates, add AndroidX, handle yarn engine checks
- Bump Kotlin from 1.4.32 to 1.5.10 (fixes JS module metadata handling for -js-legacy variants)
- Replace deprecated KotlinCompile DSL type with tasks type for 1.5.x compatibility
- Fix normalizedWaveValueToByte: .toByte() was deprecated in 1.5.0 → use .toInt().toByte()
- Use tasks.named instead of tasks.withType for metadataSourcesJar duplicate strategy
- Change KotlinCompile DSL type to tasks type for 1.5.x compatibility
- Add android.useAndroidX=true to gradle.properties
- Add --ignore-engines to yarn install via KotlinNpmInstallTask args
- Replace remaining jcenter() with mavenCentral() in ktlint.gradle.kts

+++ cli/src/main/kotlin/it/krzeminski/fsynth/SoundOutputRendering.kt
@@ -12,4 +12,4 @@ fun render8bit(song: Song, sampleRate: Int, startTime: Float): ByteArray =
  * Takes [value] of the sound wave, between -1.0 and 1.0, and returns an 8-bit sample (e.g. 0 for -1.0, 255 for 1.0).
  */
 private fun normalizedWaveValueToByte(value: Float): Byte =
-        (((value + 1.0f) / 2.0f) * 255.0f).toByte()
+        (((value + 1.0f) / 2.0f) * 255.0f).toInt().toByte()
